requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
ecdsa==0.17.0
frozenlist==1.3.0
humanize==4.1.0
idna==3.3
importlib-metadata==4.11.4
ms-cv==0.1.1
multidict==6.0.2
PSNAWP==2021.20.6
psycopg2==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
six==1.16.0
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
xbox-webapi==2.0.11
yarl==1.7.2
zipp==3.8.0
